LayoutTests/canvas/philip/tests/2d.imageData.get.order.cols-expected.txt

< [index] >
2d.imageData.get.order.cols
getImageData() returns leftmost columns first
References: 2d.pixelarray.order
Actual output:
Passed
